chore: add file for rewards history table
